% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{postClipboard2Gitterchat}
\alias{postClipboard2Gitterchat}
\title{Post clipboard text to gitter chatroom}
\usage{
postClipboard2Gitterchat()
}
\value{

}
\description{
must set .roomId in the global environment
}
\examples{
none
}
