% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sheet_write_threadMsgInfoFromLinkAddress}
\alias{sheet_write_threadMsgInfoFromLinkAddress}
\title{Extract a gitter thread info and write to google sheet}
\usage{
sheet_write_threadMsgInfoFromLinkAddress()
}
\value{

}
\description{
need to have gitter's roomId stored as .roomId in the global environment, also a .sheetname in the global environment as well.
}
\examples{
none
}
